Add awaitRestore to customEntitlementComputation version
